backend/requirements.txt
# Framework principal
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Base de datos
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.30.0
aiosqlite==0.20.0

# Validación y serialización
pydantic==2.7.1
pydantic-settings==2.2.1

# Autenticación y seguridad
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

# Configuración
python-dotenv==1.0.1

# Utilidades
pillow==10.3.0
aiofiles==23.2.1
python-slugify==8.0.4
email-validator==2.1.1

# Desarrollo y testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0